added nikhil's algorithm as well
added nikhil's suggested algorithm in the code as well, but it doesn't provide a noticeable speedup. therefore, I've added the function but I've kept the rest of the code as the same (which doesn't call this new function at all). however, i doubt if the lack of speedup is due to using an expensive datatype i.e. std::vector instead of an array with size nspins-1. however, we'll deal with that later. more or less, i'm done with the classical ising model code now.